Game Fixes (#55)
* Health Bar Edits

Edited Health bar & began working on player animation

* Low Health indicator

If the player is at 30% health or less the health bar becomes red otherwise it's set to green

* Fixed Player Animation

Player animation now works as intended. Also Reorganized animation files.

* Re_apply main character to main menu

* Visible invisibility

Player becomes red as long as they are invincible.

* Lava Level Fix

* Color Indicators

When player becomes blue, the charges shot is fully charged, when play is red he is invincible.

* Revert "Color Indicators"

This reverts commit 39ec8250cd3ab52baef3b99cb2bcf3bdecdee318.

* Color Indicators

When player becomes blue, the charges shot is fully charged, when play is red he is invincible. (Re-done)